# doc/api/api-contract.yml
# @version 1.3
# @description Risk Guard 統一 API 規範 (已同步最新重構風格)
# @project risk-guard

openapi: 3.0.3
info:
  title: Risk Guard API
  description: |
    系統核心 API 規範，包含認證模組與選單模組。
    已採用 `ResponseEntity<ApiResponse<T>>` 統一回傳風格。
    認證機制支援 JWT Header 與 HttpOnly Cookie (access_token, refresh_token)。
  version: 1.3.0

servers:
  - url: /api
    description: 專案基礎路徑 (例如：http://localhost:8080/api)

paths:
  # --- Authentication ---
  /auth/login:
    post:
      tags: [Authentication]
      summary: 使用者登入
      description: 驗證使用者憑據，成功後核發 JWT。支援以 email 或 username 登入。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登入成功，核發 JWT 並寫入 HttpOnly Cookie (access_token, refresh_token)
          headers:
            Set-Cookie:
              description: 包含 access_token 與 refresh_token
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_LoginResponse'
        '400':
          description: 參數錯誤
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'
        '401':
          description: 認證失敗 (帳號或密碼錯誤)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'

  /auth/register:
    post:
      tags: [Authentication]
      summary: 使用者註冊
      description: 建立新使用者帳號。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 註冊成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_LoginResponse'
        '400':
          description: 參數錯誤或 Email 已存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'

  /auth/refresh:
    post:
      tags: [Authentication]
      summary: 刷新 Access Token
      description: 支援從 Request Body 或 Cookie 中的 refresh_token 讀取，並核發新的 access_token。
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: 刷新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_LoginResponse'
        '401':
          description: Refresh Token 無效或已過期
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'

  /auth/me:
    get:
      tags: [Authentication]
      summary: 獲取當前用戶資訊
      description: 返回當前已登入使用者的基本資料與角色。
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '200':
          description: 成功獲取資訊
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_UserInfo'
        '401':
          description: 未登入或 Token 已失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'

  /auth/logout:
    post:
      tags: [Authentication]
      summary: 使用者登出
      description: 清除客戶端的 HttpOnly Cookies。
      responses:
        '200':
          description: 登出成功
          headers:
            Set-Cookie:
              description: 清除 access_token 與 refresh_token
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Void'

  # --- Menu ---
  /menu:
    get:
      tags: [Menu]
      summary: 獲取使用者選單
      description: 根據當前登入使用者的權限，返回經過濾後的導航選單。
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '200':
          description: 成功返回選單數據
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_MenuList'
        '401':
          description: 未登入
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'

components:
  schemas:
    LoginRequest:
      type: object
      required: [password]
      properties:
        email: { type: string, format: email, description: "優先使用 email 登入" }
        username: { type: string, description: "若無 email 則使用 username" }
        password: { type: string, format: password }

    RegisterRequest:
      type: object
      required: [email, username, password]
      properties:
        email: { type: string, format: email }
        username: { type: string, minLength: 3, maxLength: 20 }
        password: { type: string, format: password, minLength: 6 }

    RefreshTokenRequest:
      type: object
      properties:
        token: { type: string, description: "可選，若無則從 Cookie 'refresh_token' 讀取" }

    LoginResponse:
      type: object
      properties:
        accessToken: { type: string }
        user: { $ref: '#/components/schemas/UserInfo' }

    UserInfo:
      type: object
      properties:
        id: { type: integer, format: int64 }
        email: { type: string }
        name: { type: string, description: "對應後端 User.username" }
        roles: { type: array, items: { type: string } }

    MenuItem:
      type: object
      properties:
        id: { type: string }
        name: { type: string, description: "內部名稱" }
        label: { type: string, description: "顯示文字" }
        path: { type: string, description: "導航路徑" }
        icon: { type: string, description: "圖標名稱" }
        order: { type: integer, description: "排序權重" }
        meta: { type: object, description: "額外元數據" }
        children:
          type: array
          items: { $ref: '#/components/schemas/MenuItem' }

    ApiResponse_LoginResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse_Base'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/LoginResponse' }

    ApiResponse_UserInfo:
      allOf:
        - $ref: '#/components/schemas/ApiResponse_Base'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/UserInfo' }

    ApiResponse_MenuList:
      allOf:
        - $ref: '#/components/schemas/ApiResponse_Base'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/MenuItem' }

    ApiResponse_Void:
      allOf:
        - $ref: '#/components/schemas/ApiResponse_Base'
        - type: object
          properties:
            data: { type: object, nullable: true }

    ApiResponse_Error:
      allOf:
        - $ref: '#/components/schemas/ApiResponse_Base'
        - type: object
          properties:
            data: { type: object, nullable: true, example: null }

    ApiResponse_Base:
      type: object
      properties:
        code: { type: integer, example: 200 }
        message: { type: string, example: "操作成功" }
        timestamp: { type: integer, format: int64, description: "毫秒級時間戳" }

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token